chore: move test setup files
