+++ vitest.config.mts
@@ -12,7 +12,7 @@ export default defineConfig({
       exclude: ['src/utils/random.*', 'src/config.ts', 'src/deploy-command.ts', 'src/clients/**.ts', 'src/*-commands/index.ts', 'src/*-commands/**/utils.ts'],
     },
     clearMocks: true,
-    setupFiles: ['./mocks/msw/setup.ts'],
-    globalSetup: ['./mocks/database/globalSetup.ts'],
+    setupFiles: ['./test/mocks/msw/setup.ts'],
+    globalSetup: ['./test/mocks/database/globalSetup.ts'],
   },
 });
